+++ RELEASE_NOTES.md
@@ -3,4 +3,6 @@
 Adding support for the Widget/Content endpoints. See the [Widgets Tutorial](https://developer.bigcommerce.com/api-docs/store-management/widgets/tutorials/tutorial)
 for usage.
 
- - Implement the [Widgets API](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget/createwidget)
+Implemented:
+ - [Widgets API](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget/createwidget)
+ - [Widget Templates](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget-template/createwidgettemplate)